* sntp.sync() while disconnected now doesn't cause reset * debug write log does'nt crashed when input is nil

+++ src/init_wifi.lua
@@ -10,14 +10,16 @@ local print = nil
 if(settings.write_log) then
   print = function(input)
     _print(input)
-    local stat = file.stat(".log")
-    if (stat and stat.size > 3000000) then 
-      file.remove(".log")
-    end
-    log_file = file.open(".log", "a+")
-    if (log_file) then 
-      log_file:writeline(input)
-      log_file:close()
+    if(input) then
+      local stat = file.stat(".log")
+      if (stat and stat.size > 3000000) then 
+        file.remove(".log")
+      end
+      log_file = file.open(".log", "a+")
+      if (log_file) then 
+        log_file:writeline(input)
+        log_file:close()
+      end
     end
   end
   print("\n")
